+++ print_x.c
@@ -12,7 +12,7 @@ int print_x(va_list x, flag *f)
 	int length = 0;
 	unsigned long int n;
 
-	n = va_arg(x, unsigned int);
+	n = va_arg(x, unsigned long int);
 	if (n > 0)
 	{
 		if (f->hash)
